requirements.txt
agate==1.9.1
aiosqlite==0.21.0
alembic==1.12.1
annotated-types==0.7.0
anyio==4.8.0
apprise==1.9.2
asgi-lifespan==2.1.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
babel==2.17.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.6
coolname==2.2.0
croniter==2.0.7
cryptography==44.0.2
daff==1.3.46
dateparser==1.2.0
dbt-adapters==1.14.1
dbt-common==1.15.0
dbt-core==1.9.0
dbt-extractor==0.5.1
dbt-postgres==1.9.0
dbt-semantic-interfaces==0.7.4
deepdiff==7.0.1
Deprecated==1.2.18
dnspython==2.7.0
docker==6.1.3
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.11
fsspec==2025.3.0
google-auth==2.38.0
graphviz==0.20.3
greenlet==3.1.1
griffe==1.6.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.12.1
hyperframe==6.1.0
idna==3.10
importlib-metadata==6.11.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.3
jinja2-humanize-extension==0.4.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==28.1.0
leather==0.4.0
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mashumaro==3.14
mdurl==0.1.2
more-itertools==10.6.0
msgpack==1.1.0
numpy==1.26.4
networkx==3.2.1
oauthlib==3.2.2
OdooRPC==0.10.1
opentelemetry-api==1.31.0
ordered-set==4.1.0
orjson==3.10.15
packaging==24.2
pandas==2.2.0
parsedatetime==2.6
pathspec==0.12.1
pendulum==3.0.0
prefect==3.2.13
prometheus_client==0.21.1
protobuf==5.29.3
psycopg2-binary==2.9.9
pyactiveresource==2.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-slugify==8.0.4
python-socks==2.7.1
pytimeparse==1.1.8
pytz==2023.4
pytzdata==2020.1
PyYAML==6.0.2
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
shellingham==1.5.4
ShopifyAPI==12.7.0
six==1.17.0
sniffio==1.3.1
snowplow-tracker==1.1.0
SQLAlchemy==2.0.25
sqlparse==0.5.3
starlette==0.46.1
text-unidecode==1.3
time-machine==2.16.0
toml==0.10.2
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
ujson==5.10.0
urllib3==1.26.20
uv==0.6.6
uvicorn==0.34.0
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
zipp==3.21.0
tenacity==9.0.0
dotenv
Flask==2.2.2


